requirements.txt
flask==3.1.0
gunicorn==23.0.0
tensorflow-cpu==2.18.0
keras==3.13.2
numpy==1.26.4
pillow==11.1.0
h5py==3.12.1
